Correct descriptions now show in AdventureActivity
The description also has a "trim length" of 200 characters, at which it
will hide further text that can be revealed by pressing on the
ExpandableTextView.

TODO: Load a placeholder image for Adventures that don't have a chosen
image, and set a uniform height for the ImageView that contains the
images.